refactor(handlers): extract group handlers (WP5.9.5.3)
Extract group-related handlers into separate useGroupHandlers hook:
- findMemberNumber
- addFrequentPartner
- sameGroup
- handleSuggestionClick
- handleGroupSuggestionClick
- handleAddPlayerSuggestionClick
- handleGroupSelectCourt
- handleStreakAcknowledge
- handleGroupJoinWaitlist

Note: isPlayerAlreadyPlaying remains in useRegistrationHandlers as its
used by both groupHandlers and sendGroupToWaitlist core handler.

useRegistrationHandlers.js: 993 -> 751 lines (-242 lines)
groupHandlers.js: 369 lines (new)
