+++ src/registration/appHandlers/handlers/index.js
@@ -1,2 +1,3 @@
 export { useAdminHandlers } from './adminHandlers.js';
 export { useGuestHandlers } from './guestHandlers.js';
+export { useGroupHandlers } from './groupHandlers.js';
